Collision tuning
Changed places of the Monday code to work better. Stopped stuttering
when running into wall, stopped 'warping' to top when jumping at a wall.
When standing on an edge and jumping high, the velocity still brings you
through and is considered a left or right collision so forced out
instead of landing on top. Other than that collisions seem to be working
how I expect now. Also put a Samus texture on the player model for
visual appeal.